[#32] Update formula field
