+++ src/module/data/_module.mjs
@@ -1,5 +1,5 @@
-// Export Actors
 export * as ActiveEffect from "./effect/_module.mjs";
 export * as Actor from "./actor/_module.mjs";
-export * as Item from "./item/_module.mjs";
 export * as helpers from "./helpers.mjs";
+export * as Item from "./item/_module.mjs";
+export * as fields from "./fields/_module.mjs";
