+++ src/module/data/fields/_module.mjs
@@ -1,0 +1,1 @@
+export {default as FormulaField} from "./formula-field.mjs";
