+++ src/module/data/item/ability.mjs
@@ -1,5 +1,5 @@
 import {PowerRoll} from "../../helpers/rolls.mjs";
-import {FormulaField} from "../helpers.mjs";
+import FormulaField from "../fields/formula-field.mjs";
 import BaseItemModel from "./base.mjs";
 
 /**
